Move duplicates to a common file
